Bug fix when parsing expressions
Fixed huge bug where, on the last expression, it would matching it but then just ignore the returned ast::Expr. This was because of a buggy for loop in main.

Also redid the Parser::parseVariable() method since it was a bit hard to read and might've contributed to the bug.

Next i think we should deprecate Parser::consumeToken() since it's a bit hard to read in my opinion.

Lastly i updated the .gitignore to ignore temporary test input and output files.