First working (ich) version!
Changed the names of some members of TokenType to be more consistent.

Changed it to treat int literals the same as float literals to avoid error messages since int won't be supported until types are added. This is a temporary  change.

Fixed some things with the ast nodes and changed them to use double instead of float.

Implemented some binary operators.

NOTE: In cases where "a" is a variable reference and "b" is a literal, the program will either crash or give an assertion error since "a" is a pointer and "b" is a literal. Example
let a = 10.0
let b = 10.0
a == b // This works fine

10.0 == 10.0 // This works fine

let a = 10.0
a == 10.0 // This will not work since a is a pointer